Fix error when cancelling loading of wallet. Ref: DGDG-222, commit 724edd in governance-ui.
`onClose` updates the status of the wallet so the app assumes the keystore has been loaded; however, in governance-ui `onClose` is also being called when the modal is reset, hence the error. This diff simply replaces `onClose` with `onSuccess`, which is called in governance-ui when the keystore has successfully loaded.