[FEATURE] Fetch data to show on the profile page. Ref: DGDG-30.
The profile tab should be hidden when the user has not uploaded a wallet yet. Otherwise, clicking on the tab should show the user profile.

This commit adds the `getDaoConfig` function that fetches data from the info-server. The `DEFAULT_MENU` items in `<CollapsibleMenu/>` also has an additional `public` attribute which controls what tabs are shown to users who have not uploaded a wallet yet.

The following `data-digix` attributes were added for testing:
- `Profile-Address`
- `Profile-Status`
- `Profile-QuarterPoints`
- `Profile-ReputationPoints`
- `Profile-Stake`
- `Profile-ModerationRequirements` (div container element)
- `Profile-ModerationRequirements-Reputation`
- `Profile-ModerationRequirements-Stake`

The following attributes were also added, but these still use placeholder data for now.
- `Profile-QuarterParticipation`
- `Profile-Proposals`
- `Profile-DgxClaimed`
- `Profile-KycStatus`