[FEATURE] Lock DGD from Wallet page. Ref: DGDG-33
The button on the Wallet page opens the same overlay for locking DGDs. The condition for checking if user is allowed to lock DGD is updated in `getDgdBalance` of `<ConnectedWallet/>`. The buttons in the Header, Wallet Page, and Connected Wallet Overlay should all now point to `govUI.canLockDgd.show` for consistency.

Moreover, the Wallet page and Dashboard will fetch `AddressDetails` when the page loads to ensure that the stake value is always updated.

**Note**: We also fetch `AddressDetails` after locking DGD, but the `info-server` may not be updated by that time, so this is not a reliable way to refresh the data.

`data-digix` attributes were also added to several components related to Lock DGD.