[FIX] Make sure claimable DGX value is updated when rendering the Wallet page. Fix for DGDG-32.
The value for claimable DGX is not updated when we re-render the page. This is because we are not setting `state.claimableDgx` after fetching `AddressDetails` from the `info-server`. This diff also updates how we import `getContracts`, which fixes the bug where the transaction modal fails to show.

This also adds a minor correction in parsing `DaoConfig.CONFIG_MINIMUM_REPUTATION_FOR_MODERATOR`.