[IMPROVEMENT] Use `displayName` instead of wallet address in various parts of the site
Ref: DGDG-251

This updates the following to show the user's `displayName` instead of their wallet address:
- Sidebar (beside the `Welcome` message). The `data-digix` attribute is **Sidebar-DisplayName**.
- On top of the comment editor (beside the `Comment as` message). The `data-digix` attribute is **CommentEditor-DisplayName**.
- On top of every comment on a project. The `data-digix` attribute is **CommentAuthor-DisplayName**.

To show the `displayName`, simply use `renderDisplayName` from `/api/graphql-queries/users` and pass the `data-digix` attribute of the element.

Other parts of the site will be updated in the future once the endpoints for them have been updated.