[FIX] Update `state.lockedDgd` in Wallet page after locking DGD
Fix for DGDG-33.

This bug causes the unlock DGD overlay to not have the updated `MAX_AMOUNT` when the user locks DGD. This is because we are only updating the `stake` on the page after locking, but we forget to update the `lockedDgd` value.

This also fixes a minor bug in the Lock DGD overlay, where the phase name is not shown correctly.