[FIX] Disable lock DGD button in Profile page after locking DGD.
When locking DGD, we need to wait for the transaction to be confirmed by the `info-server`. Since there's no reliable way (yet) to fetch this data and update the values on the page immediately after the action, we should disable the button to prevent the user from locking more DGD until the transaction has been confirmed.

**Test Plan**:
- Load a participant user then go to the Profile page.
- Lock DGD. The snackbar should show up and the Lock DGD button should be disabled. However, the stake should remain the same.
- Check the data on the `info-server` to confirm that the transaction passed.
- Go to another tab then return to the Profile page. The stake should now be updated and the user can lock DGD again.