[FEATURE] Add KYC dashboard for KYC Officers
This allows KYC Officers (only) to go to the KYC dashboard. The dashboard displays the list of users who submitted their KYC and allows the KYC Officer to do any of the following actions:

- Filter list of users based on status
- Approve KYC
- Reject KYC

Non-KYC officers should not be able to access this page.

**Test Plan**

Use `account[2]` (the `KYC Officer` account) to access the dashboard.