[FIX] Address Details should be optional in KYC submission
Ref: DGDG-248

There is a bug where the `Next` button isn't enabled until you focus on the input for Address Details. This is because `addressDetails.hasError` was set to `false` instead of `undefined` initially, which means the input needs to run the validation test at least once for the button to be enabled. Since Address Details is optional, `hasError` should be set to `false` to make the field already valid.

This reverts the line change from commit `b24dc7`.