[REFACTOR] Refactor comments to use GraphQL queries. Ref: DGDG-301 (#97)
**Dev Notes**

`apollographql` does not have an existing API for invalidating parts of the cache. Thus, a workaround is provided in the `componentWillUnmount` method of the `<CommentThread/>` component to invalidate the comments cache. This is needed since we update the cache manually with new data whenever the user adds a comment, and that data may not necessarily be compatible with the server data. This also forces a refetch whenever we load the project page, ensuring that the data is always updated.

For more information, see the following github issue: https://github.com/apollographql/apollo-feature-requests/issues/4.

**Test Plan**

Regression test on comments. Make sure that the following still works:
- Add, delete, like, unlike comments. Make sure these changes persist when we reload the page.
- Loading more comments. Test on at least three levels of nested comments.
- Sorting comments.