[FEATURE] Add error overlay for project actions
Ref: DGDG-352, DGDG-311

This checks that the user meets the requirements needed for triggering a proposal action.

**Test Plan**

When clicking a proposal action button, the error overlay should appear if any of the following conditions are met:
- User is not KYC verified.
- User clicks during the locking phase.
- `proposal.prl` is not `true` (applies only for the **Claim Funding** action)

The error messages should stack in the overlay if the user meets more than one condition. If none of these conditions are met, the proposal action should continue as expected.

The following are the actions that should be checked for. All of these actions can be done by a participant i.e. no moderator actions are checked for errors.
- Abort
- Edit
- Finalize
- Claim Voting
- Claim Funding
- Change Funding
- Finish Milestone