[FEATURE] Add error overlay for Create Project button (#104)
Ref: DGDG-352, DGDG-311

This checks that the user meets the requirements needed for creating a project.

**Dev Notes**

The helper method `getUnmetProposalRequirements` checks the general conditions that should be met when doing any proposal action. This will be used in future commits when the error overlay will be applied to other project actions.

The method `inLockingPhase` is also added as a helper. As a consequence of this, other components that check for the current phase has been changed to use the same helper method, for easier maintenance in the future.

**Test Plan**

*Feature Test*

When clicking the `Create` button on the dashboard, the error overlay should appear if any of the following conditions are met:
- User is not KYC verified.
- User clicks during the locking phase.
- User's ETH balance is less than `DaoConfigs.CONFIG_PREPROPOSAL_COLLATERAL` (at the time of this writing, this is equal to 2 DGD).

The error messages should stack in the overlay if the user meets more than one condition.

If none of these conditions are met, the user should be redirected to the **Create Project** page.

*Regression Test*

The following should still work:
- When locking DGD, the correct stake is shown on the overlay. If in the locking phase, 1 DGD = 1 stake; otherwise, we calculate based on the current time.
- The correct phase is shown on the notification on top of the Lock DGD overlay i.e. the message `You are now locking in the ___ phase`.
- **Unlock DGD** button on the Wallet page should only be visible during the locking phase.