[FEATURE] Grant forum admin permission to view projects (#113)
Ref: DGDG-301

Forum Admin accounts should be able to view projects so they can individually ban/unban comments. Project action buttons (e.g. Abort, Edit, etc.) will not be available for these users. Since the dashboard checks the apollo cache for `currentUser`, we make sure that the data is available and updated by re-fetching the data when the wallet is loaded.

**Test Plan**
- Load `account15` and check that they can view projects and add comments.
- Regression test: Verify that `participant` and `moderator` accounts can still view projects and do project actions on them.