[FIX] Fix UI bugs and refactor rendering of milestones.
Ref: DT-83.

This diff implements minor UI fixes for the Project page and refactors how we render milestones.

For the milestone refactor, the logic for change fundings is modularized and moved to the `getChangedFundings` and `renderProposalFundings` methods in the `<Proposal/>` component, to make the logic easier to maintain. `finalReward` is also added to `dijixObject` when we create the attestation, so we can keep track of the reward changes in each project version.

The following are the UI changes:
- Add `ETH` label to milestone accordion in Project page.
- Hide comment filter when there are no comments to show.
- Show previous reward and milestone values when viewing previous versions of the project.
- Reward and milestone differences should show in the project page after the user changes funds (after the project has been approved).
  - Make sure to also test how it looks when there are no changes in the milestone or reward after changing funds.