[IMPROVEMENT] Refactor Wallet page and use address subscriptions (#134)
Ref: DGDG-300

This refactor modularizes the Wallet page by separating (and updating) its sections into the following components: `<WalletCurrencies>`, `<QuarterSummary`, `<VotingStake>`, and `<ParticipationReward>`. This should make it easier to maintain and to see which data is relevant for each section.

All components now use GraphQL queries and subscriptions to check for changes in the locked DGD and unclaimed DGX.

**Test Plan**

- Locked DGD should refresh when the user locks DGD or unlocks DGD.
- Claimable DGX should refresh when the user claims their reward.

Regression tests for the Wallet page should pass.