[FEATURE] Add countdown page for DigixDao Launch (#164)
Ref: DGDG-438

The countdown is based on `DaoInfo.startOfNextQuarter` when `currentQuarter === 0`. Once the timer ends, it should redirect to the dashboard automatically.

Note that even if the timer ends, `info-server` may take some time to update the value of `currentQuarter`. A workaround for this is setting a `HasCountdown` flag in the `govUI` redux. The flag is set to `false` once the timer ends, so we can redirect to the dashboard without having to wait for `info-server` to update.

This also hides the search bar from the site.

**Test Plan**
- Go to `dao-contracts` and run the following command: `START_IN=200 ./scripts/initiate-dao-countdown.sh`. This initializes the countdown based on the current time. You can replace `200` with any duration you prefer (note that this is in seconds).
- Restart `info-server` by running `npm run dev:force`.
- Check the site and verify that the countdown is running.
- When the timer is up, it should redirect you to the dashboard.

Note that there are no proposals to show after the countdown ends. Moreover, test accounts do not have DGDs locked, so you will need to trasfer funds via `truffle console` and lock DGDs to use the site.