[FEATURE] Add APPROVING status for KYC requests in admin
Ref: DGDG-395

Note: Depends on PR #27 for `dao-server`.

The `Approving` status happens for KYC requests that have been approved by the KYC officer but are still pending confirmation in the blockchain. The status should automatically reflect in the user's profile page, since we're already subscribed to the event. This diff fixes the bug where the admin UI breaks when approving a KYC request (since the icon for `Approving` requests is missing). This also adds the filter for such requests.

**Test Plan**
- Set the `BLOCK_CONFIRMATIONS` config on `info_server` to a number greater than zero. This will ensure that we see the `Approving` status, since the transaction won't be confirmed immediately. Make sure to restart the contracts.
- Approve a KYC request on admin. The UI shouldn't break after approving.
- The KYC status on the user's profile should be updated to `Approving`.
- Verify that there is an additional filter in the KYC dashboard for `Approving` requests, titled **Pending Confirmation**.