[FIX] Fix milestone display across project versions
Ref: DGDG-479, DGDG-424

If you add a milestone to a project, the milestone will be shown as `0` while the edit project transaction is still pending confirmation from the blockchain. Although we are using `milestoneFundings` for the values to display (see PR #147), we are still basing the number of milestones to display on `dijixObject.milestones`. This diff fixes that by using `milestoneFundings` as the basis for the correct milestones to show.

**Test Plan**
- Set `BLOCK_CONFIRMATIONS` to `2` on `info-server`. Restart your contracts.
- Create a project with one milestone. You can run the `create-special-proposal` script on `dao-server` to create more transactions and confirm yours.
- Edit the project and add another milestone. Before confirming your transaction, verify that viewing the project displays only one milestone and that there is only one version of the project.
- Confirm the transaction and view the project. It should now show two milestones. If you go to the previous version, it should correctly show the first milestone you created.