[FEATURE] Show Continue Participation modal
Ref: DGDG-249

This modal is shown when the following conditions have been met:
- User has loaded their wallet
- User was a past participant with enough DGDs locked in to continue participating in the new quarter
- Current quarter's rewards have been set (`DaoInfo.isGlobalRewardsSet`)
- User has not taken action from the modal before

We check if the user has taken action on the modal by caching the following key in localStorage: `[CONFIRM_PARTICIPATION_CACHE.value]_[currentQuarter]_[address]`. Including the `currentQuarter` will make sure the key is unique for each quarter (ensuring that the modal will always pop-up in the locking phase), while the `address` differentiates the users in case different wallets are used in the same browser.

**Test Plan**
- Ensure that the account you're using satisfies the conditions above.
- Teleport to the next quarter then load your wallet.
- The modal should pop-up and ask you if you want to lock DGD, continue with your current stake, or unlock DGD. All three options should function as intended.
- If the user has completed one of the options successfully, they should not see the modal anymore. Switching tabs or reloading your wallet will not make the modal appear.
- The user should not be able to exit the modal without choosing one of the three options.
- Moving to the next quarter (teleporting to the next locking phase) will invalidate the cache and make the modal pop-up for all users again.