[FEATURE] Add text formatting to forms (#253)
Ref: DGDG-514

The user should be able to format their input for the following:
- Project details (already existing)
- Milestone description
- Comments

Validation has been added for individual milestones in the project form. This diff also runs validations for suspicious hyperlinks.

**Test Plan**
- Run the translations script.
- Verify the above to be true.
- The formatted values should persist in the project form even if you navigate through different steps.
- The formatted values should be displayed correctly when viewing the preview (during create/edit project) and the project view page.
- An error will occur if the user inputs a link that does not start in `http://`, `https://`, or `mailto:`.

Do regression testing for banning users (user will not be able to comment and the UI for this should not be broken) and for forum admin actions.