[IMPROVEMENT] Use GraphQL query to get reputation and quarter points for comments
The reputation and quarter points for comments are not showing up in the proposal page for users who do not have a wallet loaded. Since the `dao-server` is now fetching the points for us, we can just use the graphQL query to fetch that data instead of manually doing so through the `info-server`.

**Test Plan**
- Comment on a proposal.
- The reputation and quarter points for the user should show up for logged-in and guest users.