[FIX] Fix bug in rendering PDF files (#290)
This fixes a bug introduced in PR #286. PDF files for projects are not rendering properly because the source is a url. The condition has been fixed to just check that we're not rendering images.

This also puts back the graphQL query for getting the `contentType` of ID proof files in the KYC dashboard