[FIX] Refactor likes on dashboard
Ref: DGDG-547

The following have been added to the `state` for the `<LandingPage/>` component:
- `displayNames`: object of `displayNames` for the project creators, with the `proposalId` as the key
- `proposals`: list of projects (from GraphQL fetch)
- `proposalLikes`: object of total like count for each project, with the `proposalId` as the key
- `userLikes`: list of `proposalId`s liked by the user

This diff sets these `state` objects as the source of truth, and removes the need to manually filter/search `ProposalLikes` for specific information during rendering. The logic for liking a project has been moved to this component as well.

The child components for the proposal cards are now just dumb components, which ensures that the data is always updated and in sync.

**Test Plan**
- Like projects on the dashboard
- Reload the dashboard. The projects should have the updated like status and like count.
- The same should be reflected for guest users.