[FEATURE] Show actionable status for projects
Ref: DGDG-548

If a project is actionable by the current user, then a green tag with the action phase will be shown on the `ProposalCard` in the dashboard. Otherwise, the tag will only show the project stage and will be colored red. The tags are also shown for the Project Details page.

The user can also toggle the checkbox on top to show actionable items only. Toggling is tracked by the `Toggle Actionable Projects` GA event, and it should only be visible for logged in users.

**Test Plan**
- Run the translations script: `npm run ui:generate-translations`.
- Go through the project lifecycle. For each phase, check if the tags change on the dashboard and the project details.
- Special projects should have the `special` tag.
- You should be able to toggle the showing of actionable items with the checkbox on top. Make sure that it still adheres to the stage filters e.g. if you're only showing ongoing projects, ticking the checkbox should show actionable ongoing projects only.
- Toggling should fire a GA event.