[FIX] Update behavior for Continue Participation Modal
Resolves DGDG-584. Reverts PR #354.

The Continue Participation modal will have the same conditions as before (see PR #245), but with the additional condition of `AddressDetails.isParticipant === false`.

The contract function to be callled is now `withdrawDGD(1)` instead of `confirmContinuedParticipation()`.

If the user skips visiting the site during the locking phase of a new quarter, they will still see the modal but it will only give them the option to lock more DGDs.

**Test Plan**

Setup
- Ensure that the account you're using satisfies the following conditions:
  - User has loaded their wallet
  - User was a past participant with locked DGDs >= `DaoConfig.CONFIG_MINIMUM_LOCKED_DGD`
  - `DaoInfo.isGlobalRewardsSet` is `true`
  - `AddressDetails.isParticipant` is `false`
  - User has not taken action from the modal before

Test Case #1: Locking Phase
- Teleport to the locking phase of the next quarter then load your wallet by running `npm run teleport:locking_phase` on the `dao-contracts` terminal.
- The modal should pop up and ask you if you want to lock DGDs, continue with your current stake, or unlock DGDs. All three options should function as intended.

Test Case #2: Main Phase
- Teleport to the main phase of the next quarter by running `npm run teleport:main_phase` twice.
- The modal should pop up and only give you the option to lock DGDs. It should work as intended.